slursbot/requirements.txt
requests
python-dotenv
pyodbc
pandas
PyYAML
python-dateutil
SQLAlchemy
cloudscraper
xlsxwriter